+++ lib/termfront/network/client.rb
@@ -14,13 +14,10 @@ module Termfront
       end
 
       def run
-        addr = prompt_address
-        return unless addr
-
         team_size = prompt_team_size
         return unless team_size
 
-        host, port = addr.include?(":") ? addr.split(":", 2).then { |h, p| [h, p.to_i] } : [addr, Config::PVP_PORT]
+        host, port = Config::PVP_DEFAULT_ADDRESS.split(":", 2).then { |h, p| [h, p.to_i] }
         begin
           @conn.connect(host, port)
           @conn.send_msg({ t: "queue", team_size: team_size })
@@ -44,52 +41,6 @@ module Termfront
       end
 
       private
-
-      def prompt_address
-        input = Config::PVP_DEFAULT_ADDRESS
-
-        STDIN.raw do |stdin|
-          loop do
-            rows, cols = @stdout.winsize
-            buf = TerminalOutput.begin_frame(home: true, clear: true)
-            lines = Array.new(rows) { " " * cols }
-
-            title = "PvP - Enter Server Address"
-            tc = [(cols - title.size) / 2 + 1, 1].max
-            lines[rows / 2 - 3] = TerminalOutput.fit_ansi("#{" " * (tc - 1)}\e[1;96m#{title}\e[0m", cols)
-
-            prompt = "> #{input}_"
-            pc = [(cols - prompt.size) / 2 + 1, 1].max
-            lines[rows / 2 - 1] = TerminalOutput.fit_ansi("#{" " * (pc - 1)}\e[97m> #{input}\e[5m_\e[0m", cols)
-
-            hint = "(Enter to continue, ESC to cancel)"
-            hc = [(cols - hint.size) / 2 + 1, 1].max
-            lines[rows / 2 + 1] = TerminalOutput.fit_ansi("#{" " * (hc - 1)}\e[90m#{hint}\e[0m", cols)
-
-            lines.each_with_index do |line, index|
-              buf << line
-              buf << "\r\n" if index < rows - 1
-            end
-            buf << TerminalOutput.end_frame
-            TerminalOutput.write_all(@stdout, buf)
-
-            next unless IO.select([stdin], nil, nil, Config::FRAME_DT)
-
-            begin
-              data = stdin.read_nonblock(64)
-              data.each_byte do |b|
-                case b
-                when 27 then return nil
-                when 13, 10 then return input.empty? ? Config::PVP_DEFAULT_ADDRESS : input
-                when 127, 8 then input = input[0...-1] unless input.empty?
-                when 32..126 then input << b.chr
-                end
-              end
-            rescue IO::WaitReadable
-            end
-          end
-        end
-      end
 
       def prompt_team_size
         selected = 0